refactor and append settings
