--- test/e2e/pages/utils/userSettings/positronSettingsManager.ts
@@ -1,17 +1,0 @@
-/*---------------------------------------------------------------------------------------------
- *  Copyright (C) 2025 Posit Software, PBC. All rights reserved.
- *  Licensed under the Elastic License 2.0. See LICENSE.txt for license information.
- *--------------------------------------------------------------------------------------------*/
-
-import path from 'path';
-import { UserSettingsFileManager } from '../userSettingsFileManager';
-
-export function createPositronSettingsManager(userDataDir: string): UserSettingsFileManager {
-	const settingsPath = path.join(userDataDir, 'User', 'settings.json');
-
-	return new UserSettingsFileManager(settingsPath, () => ({
-		'test': 'positron-settings',
-		'editor.fontSize': 16,
-		'workbench.colorTheme': 'Default Light+',
-	}));
-}
